Move LFR_benchmark to generators
