+++ networkx/algorithms/community/__init__.py
@@ -17,7 +17,6 @@ functions as attributes of ``community``. For example::
 """
 from networkx.algorithms.community.asyn_fluid import *
 from networkx.algorithms.community.centrality import *
-from networkx.algorithms.community.community_generators import *
 from networkx.algorithms.community.kclique import *
 from networkx.algorithms.community.kernighan_lin import *
 from networkx.algorithms.community.label_propagation import *
