+++ networkx/generators/tests/test_community.py
@@ -203,3 +203,69 @@ def test_stochastic_block_model():
     assert_equal(G.nodes, GG.nodes)
     GG = nx.stochastic_block_model(sizes, probs, seed=0, sparse=False)
     assert_equal(G.nodes, GG.nodes)
+
+
+def test_generator():
+    n = 250
+    tau1 = 3
+    tau2 = 1.5
+    mu = 0.1
+    G = nx.LFR_benchmark_graph(n, tau1, tau2, mu, average_degree=5,
+                               min_community=20, seed=10)
+    assert_equal(len(G), 250)
+    C = {frozenset(G.nodes[v]['community']) for v in G}
+    assert_true(nx.algorithms.community.is_partition(G.nodes(), C))
+
+
+@raises(nx.NetworkXError)
+def test_invalid_tau1():
+    n = 100
+    tau1 = 2
+    tau2 = 1
+    mu = 0.1
+    nx.LFR_benchmark_graph(n, tau1, tau2, mu, min_degree=2)
+
+
+@raises(nx.NetworkXError)
+def test_invalid_tau2():
+    n = 100
+    tau1 = 1
+    tau2 = 2
+    mu = 0.1
+    nx.LFR_benchmark_graph(n, tau1, tau2, mu, min_degree=2)
+
+
+@raises(nx.NetworkXError)
+def test_mu_too_large():
+    n = 100
+    tau1 = 2
+    tau2 = 2
+    mu = 1.1
+    nx.LFR_benchmark_graph(n, tau1, tau2, mu, min_degree=2)
+
+
+@raises(nx.NetworkXError)
+def test_mu_too_small():
+    n = 100
+    tau1 = 2
+    tau2 = 2
+    mu = -1
+    nx.LFR_benchmark_graph(n, tau1, tau2, mu, min_degree=2)
+
+
+@raises(nx.NetworkXError)
+def test_both_degrees_none():
+    n = 100
+    tau1 = 2
+    tau2 = 2
+    mu = -1
+    nx.LFR_benchmark_graph(n, tau1, tau2, mu)
+
+
+@raises(nx.NetworkXError)
+def test_neither_degrees_none():
+    n = 100
+    tau1 = 2
+    tau2 = 2
+    mu = -1
+    nx.LFR_benchmark_graph(n, tau1, tau2, mu, min_degree=2, average_degree=5)
