+++ networkx/tests/test_all_random_functions.py
@@ -173,7 +173,7 @@ def run_all_random_functions(seed):
     t(nx.utils.discrete_sequence, n, cdistribution=cdist, seed=seed)
     t(nx.algorithms.bipartite.random_graph, n, m, p, seed=seed)
     t(nx.algorithms.bipartite.gnmk_random_graph, n, m, k, seed=seed)
-    LFR = nx.algorithms.community.LFR_benchmark_graph
+    LFR = nx.LFR_benchmark_graph
     t(LFR, 25, 3, 1.5, 0.1, average_degree=3, min_community=10,
         seed=seed, max_community=20)
     # print("done")
